+++ src/index.ts
@@ -3,4 +3,5 @@ export * from './builder';
 export * from './flavors';
 export * from './initializers';
 export * from './interfaces';
+export * from './options';
 export * from './recipes';
