+++ src/interfaces/index.ts
@@ -1,4 +1,2 @@
-export * from './constructor';
 export * from './initializer';
 export * from './flavor';
-export * from './options';
