+++ src/options/array.options.ts
@@ -1,7 +1,7 @@
-import { Constructor } from '../constructor';
+import { Type } from '@nestjs/common';
 
 import { ArraySizeOptions, DecoratorOptions } from './decorator.options';
 
 export interface ArrayOptions extends DecoratorOptions, ArraySizeOptions {
-    type: Constructor,
+    type: Type,
 }
