+++ src/options/decorator.options.ts
@@ -1,3 +1,6 @@
+// eslint-disable-next-line @typescript-eslint/ban-types
+export type ValidateOptions = Function;
+
 export interface ArraySizeOptions {
     maxItems?: number,
     minItems?: number,
@@ -8,4 +11,5 @@ export interface DecoratorOptions {
     isArray?: boolean | ArraySizeOptions,
     nullable?: boolean,
     optional?: boolean,
+    validate?: ValidateOptions,
 }
