+++ src/options/nested.options.ts
@@ -1,6 +1,7 @@
-import { Constructor } from '../constructor';
+import { Type } from '@nestjs/common';
+
 import { DecoratorOptions } from './decorator.options';
 
 export interface NestedOptions extends DecoratorOptions {
-    type: Constructor,
+    type: Type,
 }
