+++ src/options/string.options.ts
