+++ src/recipes/is-date.recipe.ts
@@ -2,8 +2,9 @@ import { IsDate } from 'class-validator';
 
 import { TypePropertyDecorator } from '../adapters';
 import { Builder } from '../builder';
-import { DateOptions, Initializer } from '../interfaces';
-import { buildArrayPropertyDecorators } from './is-array.recipe';
+import { Initializer } from '../interfaces';
+import { DateOptions } from '../options';
+import { buildCommonPropertyDecorators } from './common.recipe';
 
 const DEFAULT_FORMAT = 'date-time';
 
@@ -19,7 +20,7 @@ export function IsDateRecipe<Options>(
         type: 'string',
         format: options?.format || DEFAULT_FORMAT,
     }, initializers).add(
-        ...buildArrayPropertyDecorators(options.isArray),
+        ...buildCommonPropertyDecorators(options),
 
         // should we validate the input string before transforming to a Date object?
